+++ spec/data/complicated/config/software/gecode.rb
@@ -43,6 +43,6 @@ build do
            "--disable-qt",
            "--disable-examples"].join(" "),
           :env => configure_env)
-  command "make -j #{max_build_jobs}", :env => { "LD_RUN_PATH" => "#{install_dir}/embedded/lib" }
+  command "make -j #{workers}", :env => { "LD_RUN_PATH" => "#{install_dir}/embedded/lib" }
   command "make install"
 end