+++ spec/data/complicated/config/software/logrotate.rb
@@ -36,6 +36,6 @@ env = {
 
 build do
   patch :source => "logrotate_basedir_override.patch", :plevel => 0
-  command "make -j #{max_build_jobs}", :env => env
+  command "make -j #{workers}", :env => env
   command "make install"
 end