+++ spec/data/complicated/config/software/nagios.rb
@@ -56,7 +56,7 @@ build do
   command "bash -c \"find . -name 'Makefile' | xargs sed -i 's:-o opscode-nagios -g opscode-nagios:-o root -g root:g'\""
 
   # build it
-  command "make -j #{max_build_jobs} all", :env => { "LD_RUN_PATH" => "#{install_dir}/embedded/lib" }
+  command "make -j #{workers} all", :env => { "LD_RUN_PATH" => "#{install_dir}/embedded/lib" }
   command "sudo make install"
   command "sudo make install-config"
   command "sudo make install-exfoliation"