+++ spec/data/complicated/config/software/ncurses.rb
@@ -120,8 +120,8 @@ build do
   cmd_array << "--with-libtool" if platform == 'aix'
   command(cmd_array.join(" "),
           :env => env)
-  command "make -j #{max_build_jobs}", :env => env
-  command "make -j #{max_build_jobs} install", :env => env
+  command "make -j #{workers}", :env => env
+  command "make -j #{workers} install", :env => env
 
   # build non-wide-character libraries
   command "make distclean"
@@ -135,12 +135,12 @@ build do
   cmd_array << "--with-libtool" if platform == 'aix'
   command(cmd_array.join(" "),
           :env => env)
-  command "make -j #{max_build_jobs}", :env => env
+  command "make -j #{workers}", :env => env
 
   # installing the non-wide libraries will also install the non-wide
   # binaries, which doesn't happen to be a problem since we don't
   # utilize the ncurses binaries in private-chef (or oss chef)
-  command "make -j #{max_build_jobs} install", :env => env
+  command "make -j #{workers} install", :env => env
 
   # Ensure embedded ncurses wins in the LD search path
   if platform == "smartos"