+++ spec/data/complicated/config/software/openresty.rb
@@ -62,6 +62,6 @@ build do
            #'--with-libatomic'
           ].join(" "), :env => env
 
-  command "make -j #{max_build_jobs}", :env => env
+  command "make -j #{workers}", :env => env
   command "make install"
 end