+++ spec/data/complicated/config/software/pkg-config.rb
@@ -61,6 +61,6 @@ paths = [ "#{install_dir}/embedded/bin/pkgconfig" ]
 
 build do
   command "./configure --prefix=#{install_dir}/embedded --disable-debug --disable-host-tool --with-internal-glib --with-pc-path=#{paths*':'}", :env => configure_env
-  command "make -j #{max_build_jobs}", :env => configure_env
-  command "make -j #{max_build_jobs} install", :env => configure_env
+  command "make -j #{workers}", :env => configure_env
+  command "make -j #{workers} install", :env => configure_env
 end