+++ spec/data/complicated/config/software/ruby.rb
@@ -157,6 +157,6 @@ build do
   end
 
   command configure_command.join(" "), :env => env
-  command "#{make_binary} -j #{max_build_jobs}", :env => env
-  command "#{make_binary} -j #{max_build_jobs} install", :env => env
+  command "#{make_binary} -j #{workers}", :env => env
+  command "#{make_binary} -j #{workers} install", :env => env
 end