+++ spec/data/complicated/config/software/xproto.rb
@@ -41,6 +41,6 @@ configure_env =
 
 build do
   command "./configure --prefix=#{install_dir}/embedded", :env => configure_env
-  command "make -j #{max_build_jobs}", :env => configure_env
-  command "make -j #{max_build_jobs} install", :env => configure_env
+  command "make -j #{workers}", :env => configure_env
+  command "make -j #{workers} install", :env => configure_env
 end