+++ spec/data/complicated/config/software/zlib.rb
@@ -62,6 +62,6 @@ configure_env =
 
 build do
   command "./configure --prefix=#{install_dir}/embedded", :env => configure_env
-  command "make -j #{max_build_jobs}"
-  command "make -j #{max_build_jobs} install"
+  command "make -j #{workers}"
+  command "make -j #{workers} install"
 end